Moving in common classes that should be used elsewhere
