+++ common/src/main/scala/uk/ac/wellcome/finatra/modules/AmazonDynamoDBModule.scala
@@ -1,11 +1,10 @@
-package uk.ac.wellcome.platform.idminter.modules
+package uk.ac.wellcome.finatra.modules
 
 import javax.inject.Singleton
 
 import com.amazonaws.services.dynamodbv2.{AmazonDynamoDB, AmazonDynamoDBClientBuilder}
 import com.google.inject.Provides
 import com.twitter.inject.TwitterModule
-import uk.ac.wellcome.finatra.modules.DynamoConfigModule
 import uk.ac.wellcome.models.aws.DynamoConfig
 
 object AmazonDynamoDBModule extends TwitterModule {
