--- id_minter/src/main/scala/uk/ac/wellcome/platform/idminter/modules/SQSReader.scala
@@ -1,35 +1,0 @@
-package uk.ac.wellcome.platform.idminter.modules
-
-import com.amazonaws.services.sqs.AmazonSQS
-import com.amazonaws.services.sqs.model.{Message, ReceiveMessageRequest}
-import com.google.inject.{Inject, Provides,Singleton}
-import com.twitter.inject.{Logging, TwitterModule}
-import uk.ac.wellcome.finatra.modules.{SQSClientModule, SQSConfigModule}
-import uk.ac.wellcome.models.aws.SQSConfig
-
-import scala.collection.JavaConversions._
-import scala.concurrent.ExecutionContext.Implicits.global
-import scala.concurrent.Future
-import scala.concurrent.duration.{Duration, _}
-
-class SQSReader @Inject()(sqsClient:AmazonSQS, sqsConfig: SQSConfig, waitTime: Duration) extends Logging {
-
-  def retrieveMessage(): Future[Option[Message]] = Future {
-      info("looking for new messages ...")
-      sqsClient.receiveMessage(
-        new ReceiveMessageRequest(sqsConfig.queueUrl)
-          .withWaitTimeSeconds(waitTime.toSeconds.toInt)
-          .withMaxNumberOfMessages(1))
-        .getMessages.headOption
-  }
-
-}
-
-object SQSReaderModule extends TwitterModule {
-    val waitime = flag[String]("aws.sqs.wait.seconds", "10", "SQS read wait time in seconds")
-
-  @Singleton
-  @Provides
-  def providesSQSReader(sqsClient:AmazonSQS, sqsConfig: SQSConfig) =
-    new SQSReader(sqsClient,sqsConfig, waitime().toInt seconds)
-}
