+++ id_minter/src/test/scala/uk/ac/wellcome/platform/idminter/modules/IdGeneratorTest.scala
@@ -1,11 +1,12 @@
 package uk.ac.wellcome.platform.idminter.modules
 
 import com.gu.scanamo.Scanamo
+import com.gu.scanamo.syntax._
 import org.scalatest.concurrent.{IntegrationPatience, ScalaFutures}
 import org.scalatest.{BeforeAndAfterEach, FunSpec, Matchers}
-import uk.ac.wellcome.models.{Identifier, UnifiedItem}
-import com.gu.scanamo.syntax._
+import uk.ac.wellcome.models.{Id, Identifier, UnifiedItem}
 import uk.ac.wellcome.platform.idminter.DynamoDBLocal
+import uk.ac.wellcome.platform.idminter.steps.IdGenerator
 
 class IdGeneratorTest extends FunSpec with DynamoDBLocal with ScalaFutures with Matchers with BeforeAndAfterEach with IntegrationPatience{
