+++ id_minter/src/test/scala/uk/ac/wellcome/platform/idminter/modules/UnifiedItemExtractorTest.scala
@@ -6,6 +6,7 @@ import org.scalatest.concurrent.{IntegrationPatience, ScalaFutures}
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.models.aws.SQSMessage
 import uk.ac.wellcome.models.{Identifier, UnifiedItem}
+import uk.ac.wellcome.platform.idminter.steps.UnifiedItemExtractor
 import uk.ac.wellcome.utils.JsonUtil
 
 class UnifiedItemExtractorTest extends FunSpec with Matchers with ScalaFutures with IntegrationPatience {
